extract ElectionStatus from page as component
